+++ NEWS.md
@@ -22,6 +22,8 @@
 
 ## Bug Fixes
 
-- `@typst_str` no longer errors during interpolation when `TypstString` isn't loaded
+- `@typst_str`
+    - No longer error during interpolation when `TypstString` isn't loaded
+    - Handle interpolation and escaped interpolation in the same manner as double quotation marks
 - `AbstractFloat` values that satisfy `isinf` and `isnan` now correspond to `calc.inf` and `calc.nan`, respectively
 - `AbstractString` and `Text` no longer escape `$`